third_party/WebKit/LayoutTests/fast/dom/Element/attr-param-typechecking-expected.txt

This test checks whether passing wrong types to setAttributeNode causes a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.setAttributeNode("style"); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNode(null); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNode(undefined); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNode(new Object); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.removeAttributeNode("style"); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.removeAttributeNode(null); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.removeAttributeNode(undefined); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.removeAttributeNode(new Object); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNodeNS("style"); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNodeNS(null); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNodeNS(undefined); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS element.setAttributeNodeNS(new Object); threw exception TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object..
PASS successfullyParsed is true

TEST COMPLETE
